Clean up CI/CD + add splash screen (#33)
- addresses issue #28
- addresses issue #31
