--- .github/workflows/pre_release_checks.yml
@@ -1,103 +1,0 @@
-name: Pre-release checks
-on:
-  pull_request:
-    branches:
-      - main    # test-deploy on PRs to main branch (which represent a new release)
-
-jobs:
-
-  check_version_tag_does_not_exist:
-    name: check version
-    runs-on: ubuntu-latest
-    steps:
-      - name: Full checkout
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0  # fetch all history and tags
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v6
-        with:
-          version: ${{ vars.UV_VERSION }}
-
-      - name: Check tag
-        run: |
-          TAG="v$( uv version --short )"  
-          if git show-ref --tags --verify --quiet "refs/tags/${TAG}"; then
-            echo "Tag ${TAG} already exists --> UNSAFE to merge to main"
-            exit -1
-          else
-            echo "Tag ${TAG} does not exist --> safe to merge to main"
-          fi
-
-  unit_test:
-    name: full unit test matrix
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.10", "3.11", "3.12", "3.13"]
-        deps-resolution: ["lowest-direct", "highest"]
-        optional-deps: ["none", "benchmarking"]
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v6
-        with:
-          version: ${{ vars.UV_VERSION }}
-
-      - name: Set up venv, Python ${{ matrix.python-version }} & dependencies
-        run:  |
-            uv python install ${{ matrix.python-version }}
-            uv venv --python ${{ matrix.python-version }}
-            source .venv/bin/activate
-            if [ "${{ matrix.optional-deps }}" = "benchmarking" ]; then
-                uv pip compile --resolution ${{ matrix.deps-resolution }} --extra benchmarking --group dev pyproject.toml
-                uv sync --extra benchmarking --dev 
-            else
-                uv pip compile --resolution ${{ matrix.deps-resolution }} --group dev pyproject.toml
-                uv sync --dev
-            fi
-
-      - name: Run tests
-        run:  |
-            if [ "${{ matrix.optional-deps }}" = "benchmarking" ]; then
-                uv run pytest ./tests -m "not requires_no_benchmarking_deps"
-            else
-                uv run pytest ./tests/counting ./tests/shared -m "not requires_benchmarking_deps"
-            fi
-
-  deploy:
-    name: test-release (test.pypi.org)
-    needs:
-      - unit_test
-      - check_version_tag_does_not_exist
-    runs-on: ubuntu-latest
-    environment: dev
-    permissions:
-      # IMPORTANT: this permission is mandatory for Trusted Publishing
-      id-token: write
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v6
-        with:
-          version: ${{ vars.UV_VERSION }}
-
-      - name: Update version in pyproject.toml
-        run: |
-          ver_dev="$(uv version --short).dev${{ github.run_number }}"
-          echo "Updating version to ${ver_dev}"
-          uv version "${ver_dev}"
-
-      - name: Build package with uv
-        run: uv build
-
-      - name: Publish package distributions to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          repository-url: https://test.pypi.org/legacy/
