+++ .github/workflows/push_to_develop.yml
@@ -1,0 +1,42 @@
+name: Push to Develop
+on:
+  push:
+    branches: ["develop"]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+          python_version: ["3.10", "3.11", "3.12", "3.13"]
+          dependency_resolution: ["lowest-direct", "highest"]
+          include_all_extra_deps: ["false", "true"]
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - uses: ./.github/actions/unit_test
+        with:
+          uv_version: ${{ vars.UV_VERSION }}
+          python_version: ${{ matrix.python_version }}
+          dependency_resolution: ${{ matrix.dependency_resolution }}
+          include_all_extra_deps: ${{ matrix.include_all_extra_deps }}
+
+
+  update-badges-and-splash:
+    needs: test
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - uses: ./.github/actions/badges_and_splash
+        with:
+          uv_version: ${{ vars.UV_VERSION }}
+          github_token: ${{ secrets.GITHUB_TOKEN }}
