--- .github/workflows/unit_tests_reduced.yml
@@ -1,46 +1,0 @@
-name: Unit Tests - Reduced Matrix
-on:
-  push:
-    branches-ignore: ['fix/*', 'main']
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.10", "3.13"]
-        deps-resolution: ["lowest-direct", "highest"]
-        optional-deps: ["none", "benchmarking"]
-        exclude:
-          - python-version: "3.10"
-            optional-deps: "none"
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Install uv
-        uses: astral-sh/setup-uv@v6
-        with:
-          version: ${{ vars.UV_VERSION }}
-
-      - name: Set up venv, Python ${{ matrix.python-version }} & dependencies
-        run:  |
-            uv python install ${{ matrix.python-version }}
-            uv venv --python ${{ matrix.python-version }}
-            source .venv/bin/activate
-            if [ "${{ matrix.optional-deps }}" = "benchmarking" ]; then
-                uv pip compile --resolution ${{ matrix.deps-resolution }} --extra benchmarking --group dev pyproject.toml
-                uv sync --extra benchmarking --dev 
-            else
-                uv pip compile --resolution ${{ matrix.deps-resolution }} --group dev pyproject.toml
-                uv sync --dev
-            fi
-
-      - name: Run tests
-        run:  |
-            if [ "${{ matrix.optional-deps }}" = "benchmarking" ]; then
-                uv run pytest ./tests -m "not requires_no_benchmarking_deps"
-            else
-                uv run pytest ./tests/counting ./tests/shared -m "not requires_benchmarking_deps"
-            fi
