--- .github/workflows/update_badges.yml
@@ -1,97 +1,0 @@
-# Simple workflow for deploying static content to GitHub Pages
-name: Update badges
-
-on:
-  # Triggered when...
-  #  - pyproject.toml changes
-  #    - upon merge to main
-  #    - or when we manually update the version in pyproject.toml  (automatic bump does not trigger this workflow)
-  #  - update after each push to develop as well
-  push:
-    paths:
-      - "pyproject.toml"
-    branches:
-      - develop
-
-  # Or manually
-  workflow_dispatch:
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
-# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
-concurrency:
-  group: "pages"
-  cancel-in-progress: false
-
-jobs:
-  update_badges_and_publish_to_pages:
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-      - name: Install uv
-        uses: astral-sh/setup-uv@v6
-        with:
-          version: ${{ vars.UV_VERSION }}
-      - name: Update Badge - Tests & Coverage
-        run: |
-          uv venv
-          source .venv/bin/activate
-          uv sync --extra benchmarking --dev --resolution highest
-          uv run --extra benchmarking pytest ./tests -m "not requires_no_benchmarking_deps" --junitxml=junit.xml --cov=./counted_float/ --cov-report=xml
-          PACKAGE_VERSION="$(uv version --short)"
-          uv run genbadge tests -i junit.xml -o "./pages/badges/v${PACKAGE_VERSION}/tests.svg"
-          uv run genbadge coverage -i coverage.xml -o "./pages/badges/v${PACKAGE_VERSION}/coverage.svg"
-      - name: Setup Pages
-        uses: actions/configure-pages@v5
-      - name: Upload artifact
-        uses: actions/upload-pages-artifact@v3
-        with:
-          path: './pages'
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: peaceiris/actions-gh-pages@v4
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: './pages'
-          publish_branch: 'gh-pages'
-          keep_files: true
-
-  update_readme_urls:
-    needs: update_badges_and_publish_to_pages
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.13'
-
-      - name: Update README URLs
-        run: |
-          python .github/scripts/update_readme_badge_urls.py "$GITHUB_WORKSPACE/.python-versions" "$GITHUB_WORKSPACE/pyproject.toml" "$GITHUB_WORKSPACE/README.md"
-          echo "----------------------------------------------------------"
-          cat "$GITHUB_WORKSPACE/README.md" | head -n 10
-          echo "..."
-          echo "----------------------------------------------------------"
-
-      - name: Commit changes
-        run: |
-          git config --local user.name "GitHub Actions"
-          git config --local user.email ""          
-          git diff --quiet || (git add "$GITHUB_WORKSPACE/README.md" && git commit -m "Update README.md badge URLs")          
-          git push
