+++ README.md
@@ -1,8 +1,9 @@
-<!--START_SECTION:badges-->
-![shields.io](https://img.shields.io/badge/python-3.10%20%7C%203.11%20%7C%203.12%20%7C%203.13-blue)
-![genbadge](https://bertpl.github.io/counted-float/badges/v0.8.2/tests.svg)
-![genbadge](https://bertpl.github.io/counted-float/badges/v0.8.2/coverage.svg)
-<!--END_SECTION:badges-->
+<!--START_SECTION:images-->
+![shields.io-python-versions](https://img.shields.io/badge/python-3.10%20%7C%203.11%20%7C%203.12%20%7C%203.13-blue)
+![genbadge-test-count](https://bertpl.github.io/counted-float/version_artifacts/v0.8.2/badge-test-count.svg)
+![genbadge-test-coverage](https://bertpl.github.io/counted-float/version_artifacts/v0.8.2/badge-coverage.svg)
+![counted_float logo](https://bertpl.github.io/counted-float/version_artifacts/v0.8.2/splash.webp)
+<!--END_SECTION:images-->
 
 # counted-float
